Update examples for core sub-package
